包头压缩支持解析l3end
**复现步骤**

* 无

**原因和解决方案**

* 新增

**影响范围**

* 无

**验证方案**

* 验证通过

**涉及分支**

* master

**检查项**

- [ ] 需要更新依赖
- [ ] 日志打印不会造成磁盘占用空间过多
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 打包成功(如果有修改Makefile)

/assign @xiangyang

/cc @buyi @yuanchao @jinjie @wangtuo @tianxiong

+++ labeler/command.go
@@ -486,7 +486,7 @@ func parseAcl(args []string) *policy.Acl {
 				fmt.Printf("invalid proto %s from %s\n", keyValue[1], args[0])
 				return nil
 			}
-			acl.Proto = uint8(proto)
+			acl.Proto = uint16(proto)
 		case "tap":
 			switch keyValue[1] {
 			case "any":
